abi: bugfix. encoding dynamic after wide static (#77)
The prior code assumed that each field's header is 32 bytes --which is
correct for elementary static types, and for dynamic types, but is
incorrect for nested static tuples or static size arrays
with static elements.

This patch leverages the fact that the type's Size field is recursively
computed so that we account for arbitrary nesting.

As a reminder, here is the definition for encoding dynamic fields:

head(X(i)) = enc(len(head(X(1))...head(X(k)) tail(X(1))...tail(X(i-1))))
tail(X(i)) = enc(X(i))