rlp: use values instead of pointers (#106)
I don't like using sync.Pool as a band aid for performance. The extra
complexity of pointers and the opaque nature of sync.Pool
makes reasoning about GC performance more difficult than it ought to be.

In the particular case of RLP decoding, the actual problem is that we
blindly decode nested structures. This wholesale decoding leads
more allocations than may be necessary. Instead, there should be an
API that allows callers to decode one "level" at a time. This reduces
computation and allocations at the cost of a little more work on the
caller's behalf.

A forthcoming pull request will contain this new API.

Revert "rlp: use sync.Pool for Items"
This reverts commit bb9da44206f460f4cb89b72f04d8c06484a4f3ab.

Revert "rlp: use pointers to rlp Items"
This reverts commit 503ae3daf4bb13facc2206e2ad5396ff61773d09.