abi2: new package (#169)
This is a new package for decoding ABI data. The main difference between
abi and abi2 is that abi2 is meant to get it's abi schema from a json
file and that json file's event inputs can be annotated with a column
field to indicate that they are to be written to a database table.

This means that the abi2 decoding function takes an ABI payload (which
can included nested tuples, arrays, etc.) and maps it onto a one or more
rows. Multiple rows are returned when an selected input is nested in an
array. If multiple inputs are selected from different arrays within the
ABI tree, then the number of rows returned is equal to the total number
of array elements. Finally, non-array based selected fields are combined
with the array based selected fields. For example:

	Tuple(
		selected(0, static)		//f1
		array(selected(1, static))	//f2
		array(selected(2, static))	//f3
	)

The number of rows would be: |f1| + |f2|
The number of columns would be: 2

Finally, a new Integration is also introduced. This means that code is
no longer requried to add new integratinos. You can simply provide an
annotated JSON ABI file.