eth/types: refactor. relocate receipts from block to tx (#221)
There are 2 major changes in this commit.

1. Remove Receipts ([]Receipt) from the Block struct

I don't see a reason to keep the receipts separate from the Transaction
since an Ethereum Tx always has a single Receipt. Keeping them separate
required extra attention to array indexing when you needed to iterate
over the transactions and the receipts simultaneously.

2. Remove dependency on the order of Block.Txs and Tx.Receipt.Logs

Instead of having an implicit index, we now have an explicit field on
the struct for a tx or log's index. This is nice because now we can
partially download a block's transactions or a transaction's logs.