jrpc2: needs based rpc method selector (#222)
Shovel will inspect your configuration to find which data you are
indexing and it will pick the best set of rpc methods to use (best
meaning the least amount of data transferred) for the task. As it
stands, a task will be limited by the most expensive integration.
The workaround here is you can create multiple eth_sources /
integration sets to split out cheap and expensive indexers.

For example, getting a receipt tx status and a tx_input is the most
expensive set of fields to index since we have to download receipts
and transaction bodies. So if you wanted to index this set of data,
you could create a source and an integration for this specific
combination. Simultaneously, you could create an eth_source/integration
that indexed erc20 transfers. This integration would only use getLogs
-- the fastest possible rpc configuration. Both of these integration
would run at the same time, however, the erc20 transfers integration
would be significantly faster than the tx_status/tx_input integration.