shovel: add websocket to rpc client for getting latest blocks
This is an optional feature. You must set ws_url -- which may be the
same value as url in some cases -- in the source config object for the
client to make use of the websocket.

If the websocket fails for any reason, Shovel will fallback on the tried
and true HTTP based JSON RPC API.

In my testing, using the websocket cut the latency time in 1/2 for
log-only indexes. This is perfectly reasonable since the bottleneck for
indexing is dependent on HTTP requests to the source api. And since the
websocket approach handles the "latest" request concurrently, the only
request we make when indexing a block is the getLogs request. Pretty
neat.